+++ ext/CTModelsJSON.jl
@@ -305,7 +305,7 @@ function CTModels.import_ocp_solution(
         "state" => _json_to_matrix(blob["state"]),
         "control" => _json_to_matrix(blob["control"]),
         "costate" => _json_to_matrix(blob["costate"]),
-        "variable" => Vector{Float64}(blob.variable),
+        "variable" => CTModels.Serialization._extract_time_vector(blob.variable),
         "path_constraints_dual" => _json_to_optional_matrix(blob["path_constraints_dual"]),
         "boundary_constraints_dual" => isnothing(bcd) ? nothing : Vector{Float64}(bcd),
         "state_constraints_lb_dual" =>
@@ -317,9 +317,9 @@ function CTModels.import_ocp_solution(
         "control_constraints_ub_dual" =>
             _json_to_optional_matrix(blob["control_constraints_ub_dual"]),
         "variable_constraints_lb_dual" =>
-            isnothing(vclbd) ? nothing : Vector{Float64}(vclbd),
+            CTModels.Serialization._extract_optional_vector(vclbd),
         "variable_constraints_ub_dual" =>
-            isnothing(vcubd) ? nothing : Vector{Float64}(vcubd),
+            CTModels.Serialization._extract_optional_vector(vcubd),
         "control_interpolation" => blob["control_interpolation"],
     )